+++ lib/ffi/gphoto2/camera_file_info.rb
@@ -1,0 +1,10 @@
+module FFI
+  module GPhoto2
+    class CameraFileInfo < FFI::Struct
+      # libgphoto2/gphoto2-filesys.h
+      layout :preview, CameraFileInfoPreview,
+             :file, CameraFileInfoFile,
+             :audio, CameraFileInfoAudio
+    end
+  end
+end