+++ lib/ffi/gphoto2/camera_file_status.rb
@@ -1,0 +1,7 @@
+module FFI
+  module GPhoto2
+    # gphoto2/gphoto2-filesys.h
+    CameraFileStatus = enum :not_downloaded,
+                            :downloaded
+  end
+end